fix(functional test): rename e2e test to functional
- rename e2e test to functional
- remove integ tests
- use eslint fix to resolve eslint issues
