+++ src/libs/tmdk/functional-tests/basic-functional/basic-functional-resources/CookieFactoryWaterTank.glb
