+++ src/libs/tmdk/jest.config.functional.js
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 module.exports = {
-  transform: {"^.+\\.ts?$": "ts-jest"},
+  transform: { "^.+\\.ts?$": "ts-jest" },
   testEnvironment: "node",
-  testRegex: "/e2e-tests/.*\\.(test|spec)?\\.(ts|tsx)$",
+  testRegex: "/functional-tests/.*\\.(test|spec)?\\.(ts|tsx)$",
   moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
   slowTestThreshold: 999999,
 };
